RENAME: style.ts 파일 이름 styles.ts로 변경
- 코딩 컨벤션을 지키기 위해 파일명 변경
